refactor: clean up coordinator card
